seperate the wf out
